feat(operationsWorkOrderActivity): implementar estructura jerárquica de actividades por responsable (#10)
- Separa la proyección plana en 4 proyecciones (Operation, WorkOrder, RootActivity, SubActivity)
- Refactoriza el use case para construir el árbol de actividades con soporte de sub-actividades anidadas
- Corrige tipo de JobTitleId de int a long en AppAttendanceDailyResponseDto
- Corrige nombre del stored procedure en OperationsWorkOrderProgressRepository

+++ Application/DTOs/Operations/OperationsAttendance/AppAttendanceDailyResponseDto.cs
@@ -79,7 +79,7 @@ namespace Application.DTOs.Operations.OperationsAttendance
         public long WorkerId { get; set; }
         public string FullName { get; set; } = string.Empty;
         public string Document { get; set; } = string.Empty;
-        public int JobTitleId { get; set; }
+        public long JobTitleId { get; set; }
         public string JobTitleDesc { get; set; } = string.Empty;
         public string? Phone { get; set; }
         public string? Email { get; set; }
